requirements.txt
pyarrow
numpy
pandas
matplotlib
seaborn
scikit-learn
tqdm
catboost
psycopg2-binary
openpyxl
gdown